cell language detection is static
